Changed type for number of MonteCarlo simulations to unsigned long long

+++ src/JSONReader.h
@@ -132,7 +132,7 @@ class JSONReader {
 			return MonteCarloSettings{
 				simulScheduler,
 				varianceReduction,
-				doc["nSimulations"].GetInt()
+				doc["nSimulations"].GetUint64()
 			};
 		}
 };
